docs: update readme
